✨ Add Support for Environment Flags `NO_COLOR` and `FORCE_COLOR` (#33)
### 🕓 Changelog

This PR adds support for the environment flags
[`NO_COLOR`](https://no-color.org) and
[`FORCE_COLOR`](https://force-color.org) to control the colour output.
Specifically, the colour output is auto-detected and can be controlled
with:

- [`NO_COLOR=true`](https://no-color.org) — disables all colours,
- [`FORCE_COLOR=true`](https://force-color.org) — forces colour output.

Only the exact value `true` is accepted to avoid accidental activation.

```console
~$ NO_COLOR=true ./safe_hashes.sh ...
~$ FORCE_COLOR=true ./safe_hashes.sh ...
```

If both are set, `NO_COLOR` takes precedence and disables all
formatting. Otherwise, colour is enabled only if output is to a
terminal, [`tput`](https://linux.die.net/man/1/tput) is available, and
the terminal supports at least the 8 standard ANSI colours.

---------

Signed-off-by: Adam Spiers <blockchain@adamspiers.org>
Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>
Co-authored-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>